# docs/swagger.yaml
basePath: /
definitions:
  handler.ChangePasswordRequest:
    properties:
      password:
        type: string
    required:
    - password
    type: object
...
  handler.DocumentListResponse:
    properties:
      list:
        items:
          type: string
        type: array
    type: object
  handler.DocumentResponse:
    properties:
      file:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handler.HuffmanResponse:
    properties:
      compressed:
        type: string
      encoding:
        additionalProperties:
          type: string
        type: object
    type: object
...
  handler.StatsResponse:
    properties:
      stat:
        items:
          $ref: '#/definitions/schema.WordStat'
        type: array
    type: object
  handler.StatusResponse:
    properties:
      status:
        type: string
    type: object
...
  schema.WordStat:
    properties:
      idf:
        type: number
      tf:
        type: number
      word:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for analyzing text documents and calculating TF-IDF statistics
  title: Text Analysis API
  version: 1.4.0
paths:
...
  /documents:
    get:
      description: Get all documents for authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.DocumentListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get list of documents
      tags:
      - documents
  /documents/{document_id}:
    delete:
      description: Delete specific document
      parameters:
      - description: Document ID
        in: path
        name: document_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete document
      tags:
      - documents
    get:
      description: Get content of specific document
      parameters:
      - description: Document ID
        in: path
        name: document_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.DocumentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get document content
      tags:
      - documents
  /documents/{document_id}/haffman:
    get:
      description: Get Huffman encoding for specific document
      parameters:
      - description: Document ID
        in: path
        name: document_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.HuffmanResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Huffman encoding
      tags:
      - documents
  /documents/{document_id}/statistics:
    get:
      description: Get statistics for specific document
      parameters:
      - description: Document ID
        in: path
        name: document_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get document statistics
      tags:
      - documents
...
  /user/{user_id}:
    delete:
      description: Delete user's account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete account
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Change user's password
      parameters:
      - description: New password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"